+ 视锥体裁剪

+++ Chapter 16 Instancing and Frustum Culling/Core/EngineTuning.cpp
@@ -362,19 +362,19 @@ __forceinline float exp2( float x ) { return pow(2.0f, x); }
 #endif
 
 ExpVar::ExpVar( const std::string& path, float val, float minExp, float maxExp, float expStepSize )
-    : NumVar(path, log2(val), minExp, maxExp, expStepSize)
+    : NumVar(path, (float)log2(val), minExp, maxExp, expStepSize)
 {
 }
 
 ExpVar& ExpVar::operator=( float val )
 {
-    m_Value = Clamp(log2(val));
+    m_Value = Clamp((float)log2(val));
     return *this;
 }
 
 ExpVar::operator float() const
 {
-    return exp2(m_Value);
+    return (float)exp2(m_Value);
 }
 
 void ExpVar::DisplayValue( TextContext& Text ) const
